argc is passed as a long by the kernel, define it as such to match what other ports do.

+++ lib/csu/mips64/md_init.h
@@ -1,4 +1,4 @@
-/* $OpenBSD: md_init.h,v 1.8 2014/12/27 20:33:47 kettenis Exp $ */
+/* $OpenBSD: md_init.h,v 1.9 2015/01/01 19:43:10 miod Exp $ */
 
 /*-
  * Copyright (c) 2001 Ross Harvey
@@ -160,7 +160,7 @@ struct kframe {
 #define	MD_START_ARGS		struct kframe *kfp, void (*cleanup)(void)
 #define	MD_START_SETUP				\
 	char	**argv, **envp;			\
-	int	argc;				\
+	long	argc;				\
 						\
 	argc = kfp->kargc;			\
 	argv = &kfp->kargv[0];			\
